# Apply currently parses the yaml object 3 times, please remove 2