@jennybuckley 2019-09-27
/lgtm
This is a really big improvement!
Also, I think we can be confident that this doesn't reintroduce the problem fixed by https://github.com/kubernetes/kubernetes/pull/75277 since the tests are all passing